feat(data-table): add data-table paths and aliases
