+++ src/components/data-table-v2/_data-table-v2-skeleton.scss
@@ -5,53 +5,4 @@
 // LICENSE file in the root directory of this source tree.
 //
 
-@import '../../globals/scss/helper-mixins';
-@import '../../globals/scss/vendor/@carbon/elements/scss/import-once/import-once';
-@import '../../globals/scss/vars';
-
-@include exports('data-table-v2-skeleton') {
-  /// @access private
-  $data-table-suffix: if(feature-flag-enabled('components-x'), '-v2', '');
-
-  .#{$prefix}--data-table#{$data-table-suffix}.#{$prefix}--skeleton {
-    th {
-      border-bottom: 1px solid $brand-01;
-
-      &:nth-child(3n + 1) {
-        width: 10%;
-      }
-
-      &:nth-child(3n + 2) {
-        width: 30%;
-      }
-
-      &:nth-child(3n + 3) {
-        width: 15%;
-      }
-    }
-
-    th span,
-    td span {
-      @include skeleton;
-      width: 75%;
-      height: 1rem;
-      display: block;
-    }
-
-    tr:hover {
-      td {
-        border-color: $ui-03;
-        background: transparent;
-
-        &:first-of-type,
-        &:last-of-type {
-          border-color: $ui-03;
-        }
-      }
-    }
-  }
-
-  .#{$prefix}--data-table-v2.#{$prefix}--skeleton .#{$prefix}--table-sort-v2 {
-    pointer-events: none;
-  }
-}
+@import '../data-table/data-table-skeleton';
