+++ src/components/data-table-v2/_data-table-v2.scss
@@ -5,9 +5,9 @@
 // LICENSE file in the root directory of this source tree.
 //
 
-@import 'data-table-v2-action';
-@import 'data-table-v2-core';
-@import 'data-table-v2-expandable';
-@import 'data-table-v2-sort';
-@import 'data-table-v2-inline-edit';
-@import 'data-table-v2-skeleton';
+@import '../data-table/data-table-action';
+@import '../data-table/data-table-core';
+@import '../data-table/data-table-expandable';
+@import '../data-table/data-table-sort';
+@import '../data-table/data-table-inline-edit';
+@import '../data-table/data-table-skeleton';
