+++ src/components/data-table/ARCHITECTURE.md
