+++ src/components/data-table/_data-table.scss
@@ -1,0 +1,13 @@
+//
+// Copyright IBM Corp. 2016, 2018
+//
+// This source code is licensed under the Apache-2.0 license found in the
+// LICENSE file in the root directory of this source tree.
+//
+
+@import 'data-table-action';
+@import 'data-table-core';
+@import 'data-table-expandable';
+@import 'data-table-sort';
+@import 'data-table-inline-edit';
+@import 'data-table-skeleton';
