+++ src/components/data-table/data-table.config.js
