+++ src/components/data-table/data-table.hbs
