+++ src/globals/scss/styles.scss
@@ -85,7 +85,7 @@ $deprecations--message: 'Deprecated code was found, this code will be removed be
 @import '../../components/link/link';
 @import '../../components/list-box/list-box';
 @import '../../components/list/list';
-@import '../../components/data-table-v2/data-table-v2';
+@import '../../components/data-table/data-table';
 @import '../../components/structured-list/structured-list';
 @import '../../components/code-snippet/code-snippet';
 @import '../../components/overflow-menu/overflow-menu';
